instrumented/special/requirements.txt
backoff==1.10.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Deprecated==1.2.13
Flask==2.0.2
googleapis-common-protos==1.54.0
grpcio==1.43.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
opentelemetry-api==1.8.0
opentelemetry-exporter-otlp==1.8.0
opentelemetry-exporter-otlp-proto-grpc==1.8.0
opentelemetry-exporter-otlp-proto-http==1.8.0
opentelemetry-instrumentation==0.27b0
opentelemetry-instrumentation-dbapi==0.27b0
opentelemetry-instrumentation-flask==0.27b0
opentelemetry-instrumentation-psycopg2==0.27b0
opentelemetry-instrumentation-requests==0.27b0
opentelemetry-instrumentation-wsgi==0.27b0
opentelemetry-proto==1.8.0
opentelemetry-sdk==1.8.0
opentelemetry-semantic-conventions==0.27b0
opentelemetry-util-http==0.27b0
protobuf==3.19.2
requests==2.27.1
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
